// gestion-backend/prisma/generated/zod/schemas/objects/NestedStringNullableFilter.schema.ts
import { z } from 'zod';
import type { Prisma } from '../../../../../generated/prisma';


export const NestedStringNullableFilterObjectSchema: z.ZodType<Prisma.NestedStringNullableFilter, z.ZodTypeDef, Prisma.NestedStringNullableFilter> = z.object({
  equals: z.string().nullish(),
  in: z.string().array().nullish(),
  notIn: z.string().array().nullish(),
  lt: z.string().optional(),
  lte: z.string().optional(),
  gt: z.string().optional(),
  gte: z.string().optional(),
  contains: z.string().optional(),
  startsWith: z.string().optional(),
  endsWith: z.string().optional(),
  search: z.string().optional(),
  not: z.union([z.string(), z.lazy(() => NestedStringNullableFilterObjectSchema)]).nullish()
}).strict();
export const NestedStringNullableFilterObjectZodSchema = z.object({
  equals: z.string().nullish(),
  in: z.string().array().nullish(),
  notIn: z.string().array().nullish(),
  lt: z.string().optional(),
  lte: z.string().optional(),
  gt: z.string().optional(),
  gte: z.string().optional(),
  contains: z.string().optional(),
  startsWith: z.string().optional(),
  endsWith: z.string().optional(),
  search: z.string().optional(),
  not: z.union([z.string(), z.lazy(() => NestedStringNullableFilterObjectSchema)]).nullish()
}).strict();